moved theme logic to custom providers
- adjusted tsconfig to react-jsx
